+++ cookbooks/arcgis-egdb/test/fixtures/cookbooks/postgresql-test/recipes/client.rb
@@ -1,0 +1,3 @@
+postgresql_client_install 'My PostgreSQL Client install' do
+    version '9.5'
+  end