+++ cookbooks/arcgis-pro/resources/pro.rb
@@ -17,19 +17,20 @@
 # limitations under the License.
 #
 
-actions :system, :install, :uninstall, :patches, :authorize
+actions :system, :unpack, :install, :uninstall, :patches, :authorize
 
+attribute :setup_archive, kind_of: String
+attribute :setups_repo, kind_of: String
 attribute :setup, kind_of: String
 attribute :product_code, kind_of: String
 attribute :install_dir, kind_of: String
 attribute :software_class, kind_of: String
 attribute :esri_license_host, kind_of: String
 attribute :authorization_type, kind_of: String
-attribute :authorization_file, :kind_of => String
-attribute :authorization_file_version, :kind_of => String
+attribute :authorization_file, kind_of: String
+attribute :authorization_file_version, kind_of: String
 attribute :portal_list, kind_of: String
-attribute :allusers, :kind_of => Integer
-
+attribute :allusers, kind_of: Integer
 
 def initialize(*args)
   super