+++ cookbooks/arcgis-insights/CHANGELOG.md
@@ -1,6 +1,9 @@
 # arcgis-insights CHANGELOG
 
 This file is used to list changes made in each version of the arcgis-insights cookbook.
+## 3.3.0
+- Added support for Insights for ArcGIS 3.0/3.1
+- Added dependency on new arcgis-repository cookbook to support global ArcGIS software repository in S3.
 
 ## 3.2.1
 - Added support for Insights for ArcGIS 2.2