+++ cookbooks/arcgis-pro/CHANGELOG.md
@@ -3,6 +3,11 @@ arcgis-pro cookbook CHANGELOG
 
 This file is used to list changes made in each version of the arcgis-pro cookbook.
 
+3.3.0
+-----
+- Added support for ArcGIS Pro 2.3/2.4.
+- Added dependency on new arcgis-repository cookbook to support global ArcGIS software repository in S3.
+
 3.2.1
 -----
 - Added support for ArcGIS Pro 2.2.